fixed a small bug (2d_raycasting)
I accidentally left the FOV constant at 60 degrees. All this causes is that the spacing between each ray isn't right and the rays end up in eachother's way. Changed it to math.pi*2, and because this is in radians it basically means our FOV is 360 degrees - exactly what I wanted.